refactor: extract chart sync and lookup maps from price-chart.tsx (#94)
