+++ frontend/src/components/price-chart.tsx
@@ -1,15 +1,15 @@
-import { useEffect, useRef, useState, useCallback, useMemo } from "react"
+import { useEffect, useRef, useCallback, useMemo } from "react"
 import {
   createChart,
   createSeriesMarkers,
   type IChartApi,
-  ColorType,
   CandlestickSeries,
   LineSeries,
   HistogramSeries,
 } from "lightweight-charts"
 import type { Price, Indicator, Annotation } from "@/lib/api"
-import { getChartTheme, useChartTheme, chartThemeOptions } from "@/lib/chart-utils"
+import { baseChartOptions, useChartTheme, chartThemeOptions } from "@/lib/chart-utils"
+import { useChartSync, type ChartEntry } from "@/lib/use-chart-sync"
 import { BandFillPrimitive } from "./chart/bollinger-band-fill"
 import { Legend, RsiLegend, MacdLegend, type LegendValues } from "./chart/chart-legends"
 
@@ -42,20 +42,9 @@ export function PriceChart({
   const mainChartRef = useRef<IChartApi | null>(null)
   const rsiChartRef = useRef<IChartApi | null>(null)
   const macdChartRef = useRef<IChartApi | null>(null)
-  const [hoverValues, setHoverValues] = useState<LegendValues | null>(null)
   const theme = useChartTheme()
 
-  // Build lookup maps
-  const closeByTime = useRef(new Map<string, number>())
-  const ohlcByTime = useRef(new Map<string, { o: number; h: number; l: number; c: number }>())
-  const sma20ByTime = useRef(new Map<string, number>())
-  const sma50ByTime = useRef(new Map<string, number>())
-  const bbUpperByTime = useRef(new Map<string, number>())
-  const bbLowerByTime = useRef(new Map<string, number>())
-  const rsiByTime = useRef(new Map<string, number>())
-  const macdByTime = useRef(new Map<string, number>())
-  const macdSignalByTime = useRef(new Map<string, number>())
-  const macdHistByTime = useRef(new Map<string, number>())
+  const { hoverValues, buildLookupMaps, syncCharts, setupSingleChartCrosshair } = useChartSync()
 
   // Compute latest values for default legend display
   const latestValues = useMemo<LegendValues>(() => {
@@ -78,162 +67,20 @@ export function PriceChart({
     }
   }, [prices, indicators, showSma20, showSma50, showBollinger, showRsiChart, showMacdChart])
 
-  const syncingRef = useRef(false)
-
-  const syncCharts = useCallback((
-    chartEntries: { chart: IChartApi; series: ReturnType<IChartApi["addSeries"]> }[],
-  ) => {
-    const charts = chartEntries.map((e) => e.chart)
-
-    // Sync visible range across all charts
-    for (const source of charts) {
-      source.timeScale().subscribeVisibleLogicalRangeChange(() => {
-        if (syncingRef.current) return
-        syncingRef.current = true
-        const timeRange = source.timeScale().getVisibleRange()
-        if (timeRange) {
-          for (const target of charts) {
-            if (target !== source) target.timeScale().setVisibleRange(timeRange)
-          }
-        }
-        syncingRef.current = false
-      })
-    }
-
-    const getValuesForTime = (key: string): LegendValues => {
-      const ohlc = ohlcByTime.current.get(key)
-      return {
-        o: ohlc?.o,
-        h: ohlc?.h,
-        l: ohlc?.l,
-        c: ohlc?.c,
-        sma20: sma20ByTime.current.get(key),
-        sma50: sma50ByTime.current.get(key),
-        bbUpper: bbUpperByTime.current.get(key),
-        bbLower: bbLowerByTime.current.get(key),
-        rsi: rsiByTime.current.get(key),
-        macd: macdByTime.current.get(key),
-        macdSignal: macdSignalByTime.current.get(key),
-        macdHist: macdHistByTime.current.get(key),
-      }
-    }
-
-    // Snap crosshair on all charts to actual data values
-    const snapCrosshair = (key: string, time: Parameters<IChartApi["setCrosshairPosition"]>[1]) => {
-      for (const entry of chartEntries) {
-        // Find the value for this chart's series from lookup maps
-        const closeVal = closeByTime.current.get(key)
-        const rsiVal = rsiByTime.current.get(key)
-        const macdVal = macdByTime.current.get(key)
-
-        // Determine which value to use based on which chart this is
-        if (entry === chartEntries[0] && closeVal !== undefined) {
-          entry.chart.setCrosshairPosition(closeVal, time, entry.series)
-        } else if (chartEntries.length > 1 && entry === chartEntries[1]) {
-          // Second chart could be RSI or MACD depending on what's enabled
-          const val = rsiVal !== undefined ? rsiVal : macdVal
-          if (val !== undefined) entry.chart.setCrosshairPosition(val, time, entry.series)
-        } else if (chartEntries.length > 2 && entry === chartEntries[2] && macdVal !== undefined) {
-          entry.chart.setCrosshairPosition(macdVal, time, entry.series)
-        }
-      }
-    }
-
-    const clearOtherCrosshairs = (source: IChartApi) => {
-      for (const chart of charts) {
-        if (chart !== source) chart.clearCrosshairPosition()
-      }
-    }
-
-    // Sync crosshair on hover + update legend for all charts
-    for (const source of charts) {
-      source.subscribeCrosshairMove((param) => {
-        if (param.time) {
-          const key = String(param.time)
-          setHoverValues(getValuesForTime(key))
-        } else {
-          setHoverValues(null)
-        }
-
-        if (syncingRef.current) return
-        syncingRef.current = true
-        if (param.time) {
-          snapCrosshair(String(param.time), param.time)
-        } else {
-          clearOtherCrosshairs(source)
-        }
-        syncingRef.current = false
-      })
-    }
-  }, [])
-
   useEffect(() => {
     if (!mainRef.current || !prices.length) return
     if (showRsiChart && !rsiRef.current) return
     if (showMacdChart && !macdRef.current) return
 
-    // Build lookup maps
-    closeByTime.current.clear()
-    ohlcByTime.current.clear()
-    sma20ByTime.current.clear()
-    sma50ByTime.current.clear()
-    bbUpperByTime.current.clear()
-    bbLowerByTime.current.clear()
-    rsiByTime.current.clear()
-    macdByTime.current.clear()
-    macdSignalByTime.current.clear()
-    macdHistByTime.current.clear()
+    buildLookupMaps(prices, indicators)
 
-    for (const p of prices) {
-      closeByTime.current.set(p.date, p.close)
-      ohlcByTime.current.set(p.date, { o: p.open, h: p.high, l: p.low, c: p.close })
-    }
-    for (const i of indicators) {
-      if (i.sma_20 !== null) sma20ByTime.current.set(i.date, i.sma_20)
-      if (i.sma_50 !== null) sma50ByTime.current.set(i.date, i.sma_50)
-      if (i.bb_upper !== null) bbUpperByTime.current.set(i.date, i.bb_upper)
-      if (i.bb_lower !== null) bbLowerByTime.current.set(i.date, i.bb_lower)
-      if (i.rsi !== null) rsiByTime.current.set(i.date, i.rsi)
-      if (i.macd !== null) macdByTime.current.set(i.date, i.macd)
-      if (i.macd_signal !== null) macdSignalByTime.current.set(i.date, i.macd_signal)
-      if (i.macd_hist !== null) macdHistByTime.current.set(i.date, i.macd_hist)
-    }
-
-    const theme = getChartTheme()
-
-    const subChartOpts = {
-      layout: {
-        background: { type: ColorType.Solid as const, color: theme.bg },
-        textColor: theme.text,
-        attributionLogo: false as const,
-      },
-      grid: {
-        vertLines: { color: theme.grid },
-        horzLines: { color: theme.grid },
-      },
-      crosshair: { mode: 0 as const },
-      handleScroll: {
-        horzTouchDrag: true,
-        vertTouchDrag: false,
-        mouseWheel: false,
-        pressedMouseMove: true,
-      },
-      handleScale: {
-        mouseWheel: true,
-        pinch: true,
-        axisPressedMouseMove: false,
-        axisDoubleClickReset: { time: true, price: false },
-      },
-    }
+    const opts = baseChartOptions(mainRef.current, 400)
 
     // Main chart — hide time axis only when sub-charts exist below
     const hideMainTimeAxis = showRsiChart || showMacdChart
     const mainChart = createChart(mainRef.current, {
-      width: mainRef.current.clientWidth,
-      height: 400,
-      ...subChartOpts,
-      rightPriceScale: { borderColor: theme.border },
-      timeScale: { borderColor: theme.border, visible: !hideMainTimeAxis },
+      ...opts,
+      timeScale: { ...opts.timeScale, visible: !hideMainTimeAxis },
     })
 
     // Main series: candle or line
@@ -244,9 +91,7 @@ export function PriceChart({
         lineWidth: 2,
         priceLineVisible: false,
       })
-      mainSeries.setData(
-        prices.map((p) => ({ time: p.date, value: p.close }))
-      )
+      mainSeries.setData(prices.map((p) => ({ time: p.date, value: p.close })))
     } else {
       mainSeries = mainChart.addSeries(CandlestickSeries, {
         upColor: "#22c55e",
@@ -303,9 +148,7 @@ export function PriceChart({
           crosshairMarkerVisible: false,
         })
         sma20.setData(
-          indicators
-            .filter((i) => i.sma_20 !== null)
-            .map((i) => ({ time: i.date, value: i.sma_20! }))
+          indicators.filter((i) => i.sma_20 !== null).map((i) => ({ time: i.date, value: i.sma_20! }))
         )
       }
 
@@ -317,9 +160,7 @@ export function PriceChart({
           crosshairMarkerVisible: false,
         })
         sma50.setData(
-          indicators
-            .filter((i) => i.sma_50 !== null)
-            .map((i) => ({ time: i.date, value: i.sma_50! }))
+          indicators.filter((i) => i.sma_50 !== null).map((i) => ({ time: i.date, value: i.sma_50! }))
         )
       }
     }
@@ -335,70 +176,47 @@ export function PriceChart({
           text: a.title.slice(0, 2),
         }))
         .sort((a, b) => (a.time < b.time ? -1 : 1))
-
       createSeriesMarkers(mainSeries, markers)
     }
 
     mainChart.timeScale().fitContent()
     mainChartRef.current = mainChart
 
-    // Collect chart entries for sync
-    const chartEntries: { chart: IChartApi; series: ReturnType<IChartApi["addSeries"]> }[] = [
-      { chart: mainChart, series: mainSeries },
-    ]
+    const chartEntries: ChartEntry[] = [{ chart: mainChart, series: mainSeries }]
     const createdCharts: IChartApi[] = [mainChart]
 
     // RSI chart
-    let rsiSeries: ReturnType<IChartApi["addSeries"]> | null = null
     if (showRsiChart && rsiRef.current) {
+      const rsiOpts = baseChartOptions(rsiRef.current, 120)
       const rsiChart = createChart(rsiRef.current, {
-        width: rsiRef.current.clientWidth,
-        height: 120,
-        ...subChartOpts,
+        ...rsiOpts,
         rightPriceScale: {
-          borderColor: theme.border,
+          ...rsiOpts.rightPriceScale,
           autoScale: false,
           scaleMargins: { top: 0.05, bottom: 0.05 },
         },
-        timeScale: { borderColor: theme.border, timeVisible: false },
       })
 
-      rsiSeries = rsiChart.addSeries(LineSeries, {
+      const rsiSeries = rsiChart.addSeries(LineSeries, {
         color: "#8b5cf6",
         lineWidth: 2,
         priceLineVisible: false,
-        autoscaleInfoProvider: () => ({
-          priceRange: { minValue: 0, maxValue: 100 },
-        }),
+        autoscaleInfoProvider: () => ({ priceRange: { minValue: 0, maxValue: 100 } }),
       })
-
       rsiSeries.setData(
-        indicators
-          .filter((i) => i.rsi !== null)
-          .map((i) => ({ time: i.date, value: i.rsi! }))
+        indicators.filter((i) => i.rsi !== null).map((i) => ({ time: i.date, value: i.rsi! }))
       )
 
       // RSI threshold lines
-      const overbought = rsiChart.addSeries(LineSeries, {
-        color: "rgba(239, 68, 68, 0.5)",
-        lineWidth: 1,
-        lineStyle: 2,
+      const thresholdOpts = {
+        lineWidth: 1 as const,
+        lineStyle: 2 as const,
         priceLineVisible: false,
         crosshairMarkerVisible: false,
-        autoscaleInfoProvider: () => ({
-          priceRange: { minValue: 0, maxValue: 100 },
-        }),
-      })
-      const oversold = rsiChart.addSeries(LineSeries, {
-        color: "rgba(34, 197, 94, 0.5)",
-        lineWidth: 1,
-        lineStyle: 2,
-        priceLineVisible: false,
-        crosshairMarkerVisible: false,
-        autoscaleInfoProvider: () => ({
-          priceRange: { minValue: 0, maxValue: 100 },
-        }),
-      })
+        autoscaleInfoProvider: () => ({ priceRange: { minValue: 0, maxValue: 100 } }),
+      }
+      const overbought = rsiChart.addSeries(LineSeries, { ...thresholdOpts, color: "rgba(239, 68, 68, 0.5)" })
+      const oversold = rsiChart.addSeries(LineSeries, { ...thresholdOpts, color: "rgba(34, 197, 94, 0.5)" })
 
       const rsiDates = indicators.filter((i) => i.rsi !== null).map((i) => i.date)
       if (rsiDates.length) {
@@ -415,23 +233,13 @@ export function PriceChart({
     }
 
     // MACD chart
-    let macdLineSeries: ReturnType<IChartApi["addSeries"]> | null = null
     if (showMacdChart && macdRef.current) {
-      const macdChart = createChart(macdRef.current, {
-        width: macdRef.current.clientWidth,
-        height: 120,
-        ...subChartOpts,
-        rightPriceScale: { borderColor: theme.border },
-        timeScale: { borderColor: theme.border, timeVisible: false },
-      })
+      const macdOpts = baseChartOptions(macdRef.current, 120)
+      const macdChart = createChart(macdRef.current, macdOpts)
 
       const macdData = indicators.filter((i) => i.macd !== null)
 
-      // MACD histogram (rendered first so lines draw on top)
-      const macdHistSeries = macdChart.addSeries(HistogramSeries, {
-        priceLineVisible: false,
-        base: 0,
-      })
+      const macdHistSeries = macdChart.addSeries(HistogramSeries, { priceLineVisible: false, base: 0 })
       macdHistSeries.setData(
         macdData.map((i) => ({
           time: i.date,
@@ -440,18 +248,14 @@ export function PriceChart({
         }))
       )
 
-      // MACD line
-      macdLineSeries = macdChart.addSeries(LineSeries, {
+      const macdLineSeries = macdChart.addSeries(LineSeries, {
         color: "#38bdf8",
         lineWidth: 2,
         priceLineVisible: false,
         crosshairMarkerVisible: false,
       })
-      macdLineSeries.setData(
-        macdData.map((i) => ({ time: i.date, value: i.macd! }))
-      )
+      macdLineSeries.setData(macdData.map((i) => ({ time: i.date, value: i.macd! })))
 
-      // Signal line
       const macdSignalSeries = macdChart.addSeries(LineSeries, {
         color: "#fb923c",
         lineWidth: 2,
@@ -459,12 +263,9 @@ export function PriceChart({
         crosshairMarkerVisible: false,
       })
       macdSignalSeries.setData(
-        indicators
-          .filter((i) => i.macd_signal !== null)
-          .map((i) => ({ time: i.date, value: i.macd_signal! }))
+        indicators.filter((i) => i.macd_signal !== null).map((i) => ({ time: i.date, value: i.macd_signal! }))
       )
 
-      // Zero line
       if (macdData.length) {
         const zeroLine = macdChart.addSeries(LineSeries, {
           color: "rgba(161, 161, 170, 0.3)",
@@ -487,26 +288,8 @@ export function PriceChart({
     // Sync all created charts
     if (chartEntries.length > 1) {
       syncCharts(chartEntries)
-    } else if (chartEntries.length === 1) {
-      // Single chart — still need crosshair legend updates
-      mainChart.subscribeCrosshairMove((param) => {
-        if (param.time) {
-          const key = String(param.time)
-          const ohlc = ohlcByTime.current.get(key)
-          setHoverValues({
-            o: ohlc?.o,
-            h: ohlc?.h,
-            l: ohlc?.l,
-            c: ohlc?.c,
-            sma20: sma20ByTime.current.get(key),
-            sma50: sma50ByTime.current.get(key),
-            bbUpper: bbUpperByTime.current.get(key),
-            bbLower: bbLowerByTime.current.get(key),
-          })
-        } else {
-          setHoverValues(null)
-        }
-      })
+    } else {
+      setupSingleChartCrosshair(mainChart)
     }
 
     const resizeObserver = new ResizeObserver((entries) => {
@@ -528,7 +311,7 @@ export function PriceChart({
       rsiChartRef.current = null
       macdChartRef.current = null
     }
-  }, [prices, indicators, annotations, syncCharts, showSma20, showSma50, showBollinger, showRsiChart, showMacdChart, chartType])
+  }, [prices, indicators, annotations, buildLookupMaps, syncCharts, setupSingleChartCrosshair, showSma20, showSma50, showBollinger, showRsiChart, showMacdChart, chartType])
 
   // Apply theme changes without recreating charts
   useEffect(() => {
@@ -544,7 +327,6 @@ export function PriceChart({
     macdChartRef.current?.timeScale().fitContent()
   }, [])
 
-  // Determine rounding class for main chart
   const mainRoundClass = !showRsiChart && !showMacdChart ? "rounded-md" : "rounded-t-md"
 
   return (
